#5514 #5515 create new users endpoint for non-admin user commands
Currently contains two superUser commands
