+++ src/main/java/edu/harvard/iq/dataverse/api/Users.java
@@ -1,0 +1,82 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package edu.harvard.iq.dataverse.api;
+
+import static edu.harvard.iq.dataverse.api.AbstractApiBean.error;
+import edu.harvard.iq.dataverse.authorization.users.AuthenticatedUser;
+import edu.harvard.iq.dataverse.engine.command.impl.ChangeUserIdentifierCommand;
+import edu.harvard.iq.dataverse.engine.command.impl.MergeInAccountCommand;
+import java.util.logging.Logger;
+import javax.ejb.Stateless;
+import javax.ws.rs.POST;
+import javax.ws.rs.Path;
+import javax.ws.rs.PathParam;
+import javax.ws.rs.core.Response;
+
+/**
+ *
+ * @author madunlap
+ */
+@Stateless
+@Path("users")
+public class Users extends AbstractApiBean{
+    
+    private static final Logger logger = Logger.getLogger(Users.class.getName());
+    
+    @POST
+    @Path("{consumedIdentifier}/mergeIntoUser/{baseIdentifier}")
+    public Response mergeInAuthenticatedUser(@PathParam("consumedIdentifier") String consumedIdentifier, @PathParam("baseIdentifier") String baseIdentifier) {
+
+        if(null == baseIdentifier || baseIdentifier.isEmpty()) {
+            return error(Response.Status.BAD_REQUEST, "Base identifier provided to change is empty.");
+        } else if(null == consumedIdentifier || consumedIdentifier.isEmpty()) {
+            return error(Response.Status.BAD_REQUEST, "Identifier to merge in is empty.");
+        }
+
+        AuthenticatedUser baseAuthenticatedUser = authSvc.getAuthenticatedUser(baseIdentifier);
+        if (baseAuthenticatedUser == null) {
+            return error(Response.Status.BAD_REQUEST, "User " + baseIdentifier + " not found in AuthenticatedUser");
+        }
+
+        AuthenticatedUser consumedAuthenticatedUser = authSvc.getAuthenticatedUser(consumedIdentifier);
+        if (consumedAuthenticatedUser == null) {
+            return error(Response.Status.BAD_REQUEST, "User " + consumedIdentifier + " not found in AuthenticatedUser");
+        }
+
+        try {
+            execCommand(new MergeInAccountCommand(createDataverseRequest(consumedAuthenticatedUser), consumedAuthenticatedUser,  baseAuthenticatedUser));
+        } catch (Exception e){
+            return error(Response.Status.BAD_REQUEST, "Error calling ChangeUserIdentifierCommand: " + e.getLocalizedMessage());
+        }
+
+        return ok("All account data for " + consumedIdentifier + " has been merged into " + baseIdentifier + " .");
+    }
+
+    @POST
+    @Path("{identifier}/changeIdentifier/{newIdentifier}")
+    public Response changeAuthenticatedUserIdentifier(@PathParam("identifier") String oldIdentifier, @PathParam("newIdentifier")  String newIdentifier) {
+
+        if(null == oldIdentifier || oldIdentifier.isEmpty()) {
+            return error(Response.Status.BAD_REQUEST, "Old identifier provided to change is empty.");
+        } else if(null == newIdentifier || newIdentifier.isEmpty()) {
+            return error(Response.Status.BAD_REQUEST, "New identifier provided to change is empty.");
+        }
+
+        AuthenticatedUser authenticatedUser = authSvc.getAuthenticatedUser(oldIdentifier);
+        if (authenticatedUser == null) {
+            return error(Response.Status.BAD_REQUEST, "User " + oldIdentifier + " not found in AuthenticatedUser");
+        }
+
+        try {
+            execCommand(new ChangeUserIdentifierCommand(createDataverseRequest(authenticatedUser), authenticatedUser,  newIdentifier));
+        } catch (Exception e){
+            return error(Response.Status.BAD_REQUEST, "Error calling ChangeUserIdentifierCommand: " + e.getLocalizedMessage());
+        }
+
+        return ok("UserIdentifier changed from " + oldIdentifier + " to " + newIdentifier);
+    }
+    
+}
